rename table component to activeColumnsTable
