+++ lib/public/components/common/activeColumnsTable/activeColumnsTable.js
@@ -104,7 +104,7 @@ const parseColumnsConfiguration = (columns, currentProfile) => {
  * @param {TableModels|null} [models] the models that handle the table state
  * @returns {vnode} Return the total view of the table to rendered
  */
-export const table = (
+export const activeColumnsTable = (
     data,
     columnsConfiguration,
     rowsConfiguration = null,
