+++ lib/public/components/common/activeColumnsTable/headers.js
