+++ lib/public/components/common/activeColumnsTable/noDataRow.js
