+++ lib/public/components/common/activeColumnsTable/rows.js
