+++ lib/public/components/common/newTable/table.js
@@ -1,6 +1,6 @@
 import spinner from '../spinner.js';
 import errorAlert from '../errorAlert.js';
-import { noDataRow } from '../table/noDataRow.js';
+import { noDataRow } from '../activeColumnsTable/noDataRow.js';
 import { h, RemoteData } from '/js/src/index.js';
 
 /**
