+++ lib/public/utilities/applyProfile.js
@@ -11,7 +11,7 @@
  * or submit itself to any jurisdiction.
  */
 
-import { profiles } from '../components/common/table/profiles.js';
+import { profiles } from '../components/common/activeColumnsTable/profiles.js';
 
 /**
  * Apply the given profile on the given subject, and return the resulting subject or null
