+++ lib/public/views/Environments/Details/EnvironmentDetailsPage.js
@@ -15,7 +15,7 @@ import { h } from '/js/src/index.js';
 import spinner from '../../../components/common/spinner.js';
 import errorAlert from '../../../components/common/errorAlert.js';
 import { formatTimestamp } from '../../../utilities/formatting/formatTimestamp.js';
-import { table } from '../../../components/common/table/table.js';
+import { activeColumnsTable } from '../../../components/common/activeColumnsTable/activeColumnsTable.js';
 import { runsActiveColumns } from '../../Runs/ActiveColumns/runsActiveColumns.js';
 import { coloredEnvironmentStatusComponent } from '../ColoredEnvironmentStatusComponent.js';
 import { frontLink } from '../../../components/common/navigation/frontLink.js';
@@ -83,8 +83,8 @@ export const EnvironmentDetailsPage = ({ envs: { detailsModel } }) => detailsMod
                     [ENVIRONMENT_DETAILS_PANELS_KEYS.LOGS]: 'Logs',
                 },
                 {
-                    [ENVIRONMENT_DETAILS_PANELS_KEYS.RUNS]: () => table(runs, runsActiveColumns, null, { profile: 'environment' }),
-                    [LHC_FILL_DETAILS_PANELS_KEYS.LOGS]: (logs) => table(logs, logsActiveColumns, null, { profile: 'embeded' }),
+                    [ENVIRONMENT_DETAILS_PANELS_KEYS.RUNS]: () => activeColumnsTable(runs, runsActiveColumns, null, { profile: 'environment' }),
+                    [LHC_FILL_DETAILS_PANELS_KEYS.LOGS]: (logs) => activeColumnsTable(logs, logsActiveColumns, null, { profile: 'embeded' }),
                 },
             ),
         ]);
