+++ lib/public/views/Flps/Overview/index.js
@@ -12,7 +12,7 @@
  */
 
 import { h } from '/js/src/index.js';
-import { table } from '../../../components/common/table/table.js';
+import { activeColumnsTable } from '../../../components/common/activeColumnsTable/activeColumnsTable.js';
 import { flpsActiveColumns } from '../ActiveColumns/flpsActiveColumns.js';
 import { paginationComponent } from '../../../components/Pagination/paginationComponent.js';
 
@@ -28,7 +28,7 @@ const flpOverview = (model) => {
         h('h2.mv2', 'Flp'),
         h('.flex-row.mv1', [
             h('.w-90', h('.w-100.flex-column.mh3', [
-                table(data, flpsActiveColumns, {
+                activeColumnsTable(data, flpsActiveColumns, {
                     callback: (entry) => ({
                         onclick: () => model.router.go(`?page=flp-detail&id=${entry.id}`, ''),
                     }),
