+++ lib/public/views/Home/Overview/index.js
@@ -14,7 +14,7 @@
 import { h } from '/js/src/index.js';
 import { logsActiveColumns } from '../../Logs/ActiveColumns/logsActiveColumns.js';
 import { runsActiveColumns } from '../../Runs/ActiveColumns/runsActiveColumns.js';
-import { table } from '../../../components/common/table/table.js';
+import { activeColumnsTable } from '../../../components/common/activeColumnsTable/activeColumnsTable.js';
 import { estimateDisplayableRowsCount } from '../../../utilities/estimateDisplayableRowsCount.js';
 
 const TABLEROW_HEIGHT = 46;
@@ -39,11 +39,11 @@ const HomeOverviewScreen = (model) => {
             style: 'padding-right: 1rem;',
         }, [
             h('h2', 'Log Entries'),
-            table(logs, logsActiveColumns, null, { profile: 'home' }),
+            activeColumnsTable(logs, logsActiveColumns, null, { profile: 'home' }),
         ]),
         h('.flex-column.w-50', [
             h('h2', 'Runs'),
-            table(runs, runsActiveColumns, null, { profile: 'home' }),
+            activeColumnsTable(runs, runsActiveColumns, null, { profile: 'home' }),
         ]),
     ]);
 };
