+++ lib/public/views/LhcFills/Detail/lhcFillDetailsComponent.js
@@ -13,7 +13,7 @@
 
 import { h, iconWarning } from '/js/src/index.js';
 import spinner from '../../../components/common/spinner.js';
-import { table } from '../../../components/common/table/table.js';
+import { activeColumnsTable } from '../../../components/common/activeColumnsTable/activeColumnsTable.js';
 import errorAlert from '../../../components/common/errorAlert.js';
 import { runsActiveColumns } from '../../Runs/ActiveColumns/runsActiveColumns.js';
 import { commonLhcFillDisplayConfiguration } from './commonLhcFillDisplayConfiguration.js';
@@ -218,10 +218,10 @@ export const lhcFillDetailsComponent = (detailsModel) => detailsModel.lhcFillDet
                                     ]),
                                 ]),
                             ]),
-                            h('#runs', table(runs, runsActiveColumns, null, { profile: 'lhcFill' })),
+                            h('#runs', activeColumnsTable(runs, runsActiveColumns, null, { profile: 'lhcFill' })),
                         ];
                     },
-                    [LHC_FILL_DETAILS_PANELS_KEYS.LOGS]: (logs) => table(logs, logsActiveColumns, null, { profile: 'embeded' }),
+                    [LHC_FILL_DETAILS_PANELS_KEYS.LOGS]: (logs) => activeColumnsTable(logs, logsActiveColumns, null, { profile: 'embeded' }),
                 },
             ),
         ]);
