+++ lib/public/views/LhcFills/Overview/index.js
@@ -13,7 +13,7 @@
  */
 
 import { h } from '/js/src/index.js';
-import { table } from '../../../components/common/table/table.js';
+import { activeColumnsTable } from '../../../components/common/activeColumnsTable/activeColumnsTable.js';
 import { lhcFillsActiveColumns } from '../ActiveColumns/lhcFillsActiveColumns.js';
 import { estimateDisplayableRowsCount } from '../../../utilities/estimateDisplayableRowsCount.js';
 import { paginationComponent } from '../../../components/Pagination/paginationComponent.js';
@@ -51,7 +51,7 @@ const showLhcFillsTable = (lhcFillsOverviewModel) => {
             [frontLink(h('button.btn.btn-primary', 'Statistics'), 'statistics')],
         ),
         h('.w-100.flex-column', [
-            table(lhcFillsOverviewModel.lhcFills, lhcFillsActiveColumns, { classes: '.table-sm' }),
+            activeColumnsTable(lhcFillsOverviewModel.lhcFills, lhcFillsActiveColumns, { classes: '.table-sm' }),
             paginationComponent(lhcFillsOverviewModel.pagination),
         ]),
     ];
