+++ lib/public/views/Logs/ActiveColumns/logsActiveColumns.js
@@ -22,7 +22,7 @@ import { frontLink } from '../../../components/common/navigation/frontLink.js';
 import { frontLinks } from '../../../components/common/navigation/frontLinks.js';
 import { tagFilter } from '../../../components/Filters/common/filters/tagFilter.js';
 import { formatRunsList } from '../../Runs/format/formatRunsList.js';
-import { profiles } from '../../../components/common/table/profiles.js';
+import { profiles } from '../../../components/common/activeColumnsTable/profiles.js';
 import { textFilter } from '../../../components/Filters/common/filters/textFilter.js';
 import { environmentFilter } from '../../../components/Filters/LogsFilter/environments.js';
 import { formatLhcFillsList } from '../../LhcFills/format/formatLhcFillsList.js';
