+++ lib/public/views/Logs/Logs.js
@@ -13,7 +13,7 @@
 
 import { Observable, RemoteData } from '/js/src/index.js';
 import { TagFilterModel } from '../../components/Filters/common/TagFilterModel.js';
-import { SortModel } from '../../components/common/table/SortModel.js';
+import { SortModel } from '../../components/common/activeColumnsTable/SortModel.js';
 import { debounce } from '../../utilities/debounce.js';
 import { FilterInputModel } from '../../components/Filters/common/filters/FilterInputModel.js';
 import { LogCreationModel } from './Create/LogCreationModel.js';
