+++ lib/public/views/Logs/Overview/index.js
@@ -12,7 +12,7 @@
  */
 
 import { h } from '/js/src/index.js';
-import { table } from '../../../components/common/table/table.js';
+import { activeColumnsTable } from '../../../components/common/activeColumnsTable/activeColumnsTable.js';
 import { logsActiveColumns } from '../ActiveColumns/logsActiveColumns.js';
 import { estimateDisplayableRowsCount } from '../../../utilities/estimateDisplayableRowsCount.js';
 import { paginationComponent } from '../../../components/Pagination/paginationComponent.js';
@@ -48,7 +48,7 @@ const showLogsTable = (model, logs) => {
             actionButtons(),
         ]),
         h('.w-100.flex-column', [
-            table(logs, logsActiveColumns, null, null, { sort: model.logs.overviewSortModel }),
+            activeColumnsTable(logs, logsActiveColumns, null, null, { sort: model.logs.overviewSortModel }),
             paginationComponent(model.logs.pagination),
         ]),
     ];
