+++ lib/public/views/Runs/ActiveColumns/runsActiveColumns.js
@@ -35,7 +35,7 @@ import triggerValueFilter from '../../../components/Filters/RunsFilter/triggerVa
 import lhcPeriodsFilter from '../../../components/Filters/RunsFilter/lhcPeriod.js';
 import { formatRunType } from '../../../utilities/formatting/formatRunType.js';
 import definitionFilter from '../../../components/Filters/RunsFilter/definitionFilter.js';
-import { profiles } from '../../../components/common/table/profiles.js';
+import { profiles } from '../../../components/common/activeColumnsTable/profiles.js';
 import { formatDuration } from '../../../utilities/formatting/formatDuration.mjs';
 import { formatRunStart } from '../format/formatRunStart.js';
 import { formatRunEnd } from '../format/formatRunEnd.js';
