+++ lib/public/views/Runs/Details/runDetailsComponent.js
@@ -16,7 +16,7 @@ import { lhcFillPanel } from './lhcFillPanel.js';
 import { editTagsPanel } from './editTagsPanel.js';
 import { tabbedPanelComponent } from '../../../components/TabbedPanel/tabbedPanelComponent.js';
 import { RUN_DETAILS_PANELS_KEYS } from './RunDetailsModel.js';
-import { table } from '../../../components/common/table/table.js';
+import { activeColumnsTable } from '../../../components/common/activeColumnsTable/activeColumnsTable.js';
 import { logsActiveColumns } from '../../Logs/ActiveColumns/logsActiveColumns.js';
 import { flpsActiveColumns } from '../../Flps/ActiveColumns/flpsActiveColumns.js';
 import spinner from '../../../components/common/spinner.js';
@@ -109,8 +109,8 @@ export const runDetailsComponent = (runDetailsModel, router) => runDetailsModel.
                     [RUN_DETAILS_PANELS_KEYS.DPL_PROCESSES]: 'QC/PDP tasks',
                 },
                 {
-                    [RUN_DETAILS_PANELS_KEYS.LOGS]: (dataLogs) => table(dataLogs, logsActiveColumns, null, { profile: 'embeded' }),
-                    [RUN_DETAILS_PANELS_KEYS.FLPS]: (dataFlps) => table(dataFlps, flpsActiveColumns, {
+                    [RUN_DETAILS_PANELS_KEYS.LOGS]: (dataLogs) => activeColumnsTable(dataLogs, logsActiveColumns, null, { profile: 'embeded' }),
+                    [RUN_DETAILS_PANELS_KEYS.FLPS]: (dataFlps) => activeColumnsTable(dataFlps, flpsActiveColumns, {
                         callback: (entry) => ({
                             style: {
                                 cursor: 'pointer',
