+++ lib/public/views/Runs/Overview/index.js
@@ -12,7 +12,7 @@
  */
 
 import { h } from '/js/src/index.js';
-import { table } from '../../../components/common/table/table.js';
+import { activeColumnsTable } from '../../../components/common/activeColumnsTable/activeColumnsTable.js';
 import { runsActiveColumns } from '../ActiveColumns/runsActiveColumns.js';
 import { exportRunsModal } from './exportRunsModal.js';
 import { estimateDisplayableRowsCount } from '../../../utilities/estimateDisplayableRowsCount.js';
@@ -48,7 +48,7 @@ const showRunsTable = (model, runs) => {
             exportRunsButton(model),
         ]),
         h('.flex-column.w-100', [
-            table(runs, runsActiveColumns),
+            activeColumnsTable(runs, runsActiveColumns),
             paginationComponent(model.runs.pagination),
         ]),
     ];
