+++ lib/public/views/Subsystems/Details/index.js
@@ -13,7 +13,7 @@
 
 import { h } from '/js/src/index.js';
 import spinner from '../../../components/common/spinner.js';
-import { table } from '../../../components/common/table/table.js';
+import { activeColumnsTable } from '../../../components/common/activeColumnsTable/activeColumnsTable.js';
 import targetURL from '../../../utilities/targetURL.js';
 import { logsActiveColumns } from '../../Logs/ActiveColumns/logsActiveColumns.js';
 import { tabLink } from '../../../components/common/navigation/tabLink.js';
@@ -48,7 +48,7 @@ const subsystemDetails = (model) => {
             },
             logs: {
                 name: 'Logs with this subsystem',
-                content: table(dataLogs, logsActiveColumns, {
+                content: activeColumnsTable(dataLogs, logsActiveColumns, {
                     callback: (entry) => ({
                         style: {
                             cursor: 'pointer',
