+++ lib/public/views/Subsystems/Overview/index.js
@@ -12,7 +12,7 @@
  */
 
 import { h } from '/js/src/index.js';
-import { table } from '../../../components/common/table/table.js';
+import { activeColumnsTable } from '../../../components/common/activeColumnsTable/activeColumnsTable.js';
 
 const ACTIVE_COLUMNS = {
     id: {
@@ -36,7 +36,7 @@ const ACTIVE_COLUMNS = {
  */
 const subsystemOverview = (model) => [
     h('h2', { onremove: () => model.subsystems.clearSubsystems() }, 'Subsystems'),
-    table(model.subsystems.getSubsystems(), ACTIVE_COLUMNS, {
+    activeColumnsTable(model.subsystems.getSubsystems(), ACTIVE_COLUMNS, {
         callback: (entry) => ({
             onclick: () => model.router.go(`?page=subsystem-detail&id=${entry.id}`),
         }),
