+++ lib/public/views/Tags/Details/tagDetailsComponent.js
@@ -14,7 +14,7 @@ import spinner from '../../../components/common/spinner.js';
 import { TAG_DETAILS_PANELS_KEYS } from './TagDetailsModel.js';
 import { h, sessionService } from '/js/src/index.js';
 import { tagDetail } from '../../../components/tag/tagDetail.js';
-import { table } from '../../../components/common/table/table.js';
+import { activeColumnsTable } from '../../../components/common/activeColumnsTable/activeColumnsTable.js';
 import { logsActiveColumns } from '../../Logs/ActiveColumns/logsActiveColumns.js';
 import errorAlert from '../../../components/common/errorAlert.js';
 import { BkpRoles } from '../../../domain/enums/BkpRoles.js';
@@ -74,7 +74,7 @@ export const tagDetailsComponent = (tagDetailsModel) => tagDetailsModel.tagDetai
             tagDetailsModel.tabbedPanelModel,
             { [TAG_DETAILS_PANELS_KEYS.LOGS]: 'Logs with this tag' },
             {
-                [TAG_DETAILS_PANELS_KEYS.LOGS]: (currentPanelData) => table(
+                [TAG_DETAILS_PANELS_KEYS.LOGS]: (currentPanelData) => activeColumnsTable(
                     currentPanelData,
                     logsActiveColumns,
                     null,
