+++ lib/public/views/Tags/Overview/index.js
@@ -12,7 +12,7 @@
  */
 
 import { h } from '/js/src/index.js';
-import { table } from '../../../components/common/table/table.js';
+import { activeColumnsTable } from '../../../components/common/activeColumnsTable/activeColumnsTable.js';
 import { tagsActiveColumns } from '../ActiveColumns/tagsActiveColumns.js';
 
 /**
@@ -43,7 +43,7 @@ const showTagsTable = (model, tags) => [
     ]),
     h('.flex-row', [
         h('.w-100.flex-column', [
-            table(tags, tagsActiveColumns, {
+            activeColumnsTable(tags, tagsActiveColumns, {
                 callback: (entry) => ({
                     onclick: () => model.router.go(`?page=tag-detail&id=${entry.id}`),
                 }),
