+++ lib/public/views/lhcPeriods/Overview/LhcPeriodsOverviewModel.js
@@ -14,7 +14,7 @@
 import { Observable, RemoteData } from '/js/src/index.js';
 import { getRemoteDataSlice } from '../../../utilities/fetch/getRemoteDataSlice.js';
 import { PaginationModel } from '../../../components/Pagination/PaginationModel.js';
-import { SortModel } from '../../../components/common/table/SortModel.js';
+import { SortModel } from '../../../components/common/activeColumnsTable/SortModel.js';
 import { TextTokensFilterModel } from '../../../components/Filters/common/filters/TextTokensFilterModel.js';
 import { buildUrl } from '../../../utilities/fetch/buildUrl.js';
