+++ lib/public/views/lhcPeriods/Overview/LhcPeriodsOverviewPage.js
@@ -13,7 +13,7 @@
  */
 
 import { h } from '/js/src/index.js';
-import { table } from '../../../components/common/table/table.js';
+import { activeColumnsTable } from '../../../components/common/activeColumnsTable/activeColumnsTable.js';
 import { lhcPeriodsActiveColumns } from '../ActiveColumns/lhcPeriodsActiveColumns.js';
 import { paginationComponent } from '../../../components/Pagination/paginationComponent.js';
 import { filtersPanelPopover } from '../../../components/Filters/common/filtersPanelPopover.js';
@@ -28,7 +28,7 @@ export const LhcPeriodsOverviewPage = ({ lhcPeriods: { overviewModel: lhcPeriods
 }, [
     h('.flex-row.flex-grow.items-start', filtersPanelPopover(lhcPeriodsOverviewModel, lhcPeriodsActiveColumns)),
     h('.w-100.flex-column', [
-        table(
+        activeColumnsTable(
             lhcPeriodsOverviewModel.lhcPeriods,
             lhcPeriodsActiveColumns,
             { classes: '.table-sm' },
